feat(ToggleValue): make component more general and rename
This renames to ToggleValue, fully decoupling from the
ConfigurationMap components. This will allow this component
to be reused in many different situations in the UI in the future.

Closes DCOS-19132.
