+++ src/styles/components/configuration-map/styles.less
@@ -111,11 +111,6 @@
       }
     }
 
-    &-toggle-value {
-      border-bottom: 1px dotted @configuration-map-border-color;
-      cursor: pointer;
-    }
-
     &-table {
 
       &-label {
