+++ src/styles/components/toggle-value/styles.less
@@ -1,0 +1,7 @@
+& when (@toggle-value-enabled) {
+
+  .toggle-value {
+    border-bottom: 1px dotted @toggle-value-border-color;
+    cursor: pointer;
+  }
+}
