+++ src/styles/components/toggle-value/variables.less
@@ -1,0 +1,3 @@
+@toggle-value-enabled: true;
+
+@toggle-value-border-color: @neutral;
