+++ src/styles/index.less
@@ -450,6 +450,11 @@
 @import 'components/toggle-button/variables.less';
 @import 'components/toggle-button/styles.less';
 
+// Toggle Value
+
+@import 'components/toggle-value/variables.less';
+@import 'components/toggle-value/styles.less';
+
 // Tooltip
 
 @import 'components/tooltip/variables.less';
